+++ src/libs/ascent/runtimes/flow_filters/ascent_runtime_vtkh_filters.hpp
@@ -490,6 +490,18 @@ public:
 };
 
 
+//-----------------------------------------------------------------------------
+class ASCENT_API VTKHMIR : public ::flow::Filter
+{
+public:
+    VTKHMIR();
+    virtual ~VTKHMIR();
+
+    virtual void   declare_interface(conduit::Node &i);
+    virtual bool   verify_params(const conduit::Node &params,
+                                 conduit::Node &info);
+    virtual void   execute();
+};
 
 
 };